requirements.txt
flask==3.0.3
flask-cors==4.0.1
ultralytics
pillow
firebase-admin
reportlab
python-dotenv
requests
